Fixed ChangeCharacterEvent.hxc
I added a workaround for using the `Type` class' `resolveClass` method.  Because `Type` got blacklisted by Eric, you can no longer use it directly, BUT if you look at the `openfl.utils.Assets` class, you will see that it has a method with the same name, which just uses `Type`'s method, so you can use that instead of using `Type` directly.

I also added a check for if a character is null when getting the default characters because it would raise errors for songs which remove some of the characters from the stage.