Fixed ChangeCharacterEvent.hxc again somehow
So the FNF devs blocked the previous workaround I used, and I thought that the only way would be to re-write this function without using `Reflect` or `resolveClass` (which I didn't plan to do), BUT THEN I FOUND ANOTHER WORKAROUND FOR USING `resolveClass` SOMEHOW?!

So apparently, they have this class in the code called `FunkinTypeResolver`, which you can create an instance of to use `resolveClass` and `resolveEnum`, and it was created to give `haxe.Unserializer` a custom type resolvers which returns `null` if the class requested is `Dynamic`. I feel like the devs will find this workaround sooner or later, so I recommend you try to write a version of this function that doesn't use `resolveClass` or `Reflect`, but for now, this workaround works on 0.5.0 just fine.

I also removed the `static` keyword from behind `BF`, `DAD`, and `GF`, because I'm pretty sure it doesn't do anything in Hscript.